refactor(runtime): 拆分 runtimeStatus 至 types 和 utils；清理 display license 头
